feat: tests helper for mock events
