+++ tests/jest.config.ts
@@ -3,9 +3,9 @@ import type { Config } from '@jest/types';
 const config: Config.InitialOptions = {
   preset: 'ts-jest', // Use ts-jest preset for TypeScript
   testEnvironment: 'node', // Test environment is Node.js
-  testMatch: ['**/*.test.tsx'], // Match test files with .test.ts extension
+  testMatch: ['**/*.test.ts'], // Match test files with .test.ts extension
   transform: {
-    '^.+\\.tsx?$': 'ts-jest',
+    '^.+\\.ts?$': 'ts-jest',
   },
   rootDir: '../',
   moduleNameMapper: {
